# internal/contract/openapi.json
{
 "openapi": "3.1.0",
...
 "info": {
  "title": "Townsquare API",
  "version": "1.0.0",
  "description": "Schedule WhatsApp posts to groups, communities, channels and Status. Everything the web app can do. Read the agent guide at /api/v1/guide.md first. All changes are recorded in history and can be undone."
 },
 "servers": [
  {
   "url": "/api/v1"
  }
 ],
 "security": [
  {
   "bearer": []
  }
 ],
...
 "paths": {
  "/status": {
   "get": {
    "tags": [
     "Status"
    ],
    "summary": "Get status",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Status"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    },
    "description": "Connection, safe mode, today's count, time zone, and your key's scope. Call this first.",
    "operationId": "get_status",
    "x-cli": {
     "name": "status get",
     "effects": "read_only",
     "examples": [
      "townsquare status get",
      "townsquare status get --fields connected,safe_mode"
     ],
     "output_kind": "data",
     "cardinality": "single"
    }
   }
  },
  "/settings": {
   "get": {
    "tags": [
     "Settings"
    ],
    "summary": "Get settings",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Settings"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    },
    "operationId": "get_settings",
    "x-cli": {
     "name": "settings get",
     "effects": "read_only",
     "examples": [
      "townsquare settings get --fields timezone,quiet_start,quiet_end"
     ],
     "output_kind": "data",
     "cardinality": "single"
    }
   },
   "patch": {
    "tags": [
     "Settings"
    ],
    "summary": "Update settings",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WriteResult"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Settings"
       }
      }
     }
    },
    "operationId": "update_settings",
    "x-cli": {
     "name": "settings update",
     "effects": "idempotent",
     "examples": [
      "townsquare settings update --quiet-start 22:00 --quiet-end 06:30",
      "townsquare settings update --safe-mode 1"
     ],
     "output_kind": "data",
     "cardinality": "single"
    }
   }
  },
  "/targets": {
   "get": {
    "tags": [
     "Targets"
    ],
    "summary": "List targets",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Target"
         }
        }
       }
      },
      "headers": {
       "X-Total-Count": {
        "description": "Matches before limit",
        "schema": {
         "type": "integer"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    },
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Name contains"
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string",
        "enum": [
         "self",
         "status",
         "announce",
         "group",
         "channel",
         "community"
        ]
       }
      },
      "description": "Comma list"
     },
     {
      "name": "client_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "allowed",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "true/false"
     },
     {
      "name": "can_send",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "true to hide admin-only chats"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer"
      }
     },
     {
      "name": "include_gone",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     }
    ],
    "operationId": "list_targets",
    "x-cli": {
     "name": "targets list",
     "effects": "read_only",
     "examples": [
      "townsquare targets list --kind group,announce --allowed true --limit 20",
      "townsquare targets list --q isla --fields jid,name,allowed"
     ],
     "output_kind": "data",
     "cardinality": "unbounded",
     "paginated": true
    }
   }
  },
  "/targets/refresh": {
   "post": {
    "tags": [
     "Targets"
    ],
    "summary": "Refresh targets from WhatsApp",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Target"
         }
        }
       }
      }
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "503": {
      "$ref": "#/components/responses/E503"
     }
    },
    "operationId": "refresh_targets_from_whatsapp",
    "x-cli": {
     "name": "targets refresh",
     "effects": "idempotent",
     "examples": [
      "townsquare targets refresh --count"
     ],
     "output_kind": "data",
     "cardinality": "unbounded"
    }
   }
  },
  "/targets/{jid}": {
   "get": {
    "tags": [
     "Targets"
    ],
    "summary": "Get target",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Target"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    },
    "parameters": [
     {
      "name": "jid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "operationId": "get_target",
    "x-cli": {
     "name": "targets get",
     "effects": "read_only",
     "examples": [
      "townsquare targets get 120363000000000102@g.us"
     ],
     "output_kind": "data",
     "cardinality": "single"
    }
   },
   "patch": {
    "tags": [
     "Targets"
    ],
    "summary": "Update target",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WriteResult"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    },
    "description": "Set client (null clears), star, or allowlist (admin).",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TargetPatch"
       }
      }
     }
    },
    "parameters": [
     {
      "name": "jid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "operationId": "update_target",
    "x-cli": {
     "name": "targets update",
     "effects": "idempotent",
     "examples": [
      "townsquare targets update 120363000000000102@g.us --client-id 2",
      "townsquare targets update 120363000000000102@g.us --allowed true --dry-run"
     ],
     "output_kind": "data",
     "cardinality": "single"
    }
   }
  },
...
  "/sets": {
   "get": {
    "tags": [
     "Sets"
    ],
    "summary": "List group sets",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Set"
         }
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    },
    "operationId": "list_group_sets",
    "x-cli": {
     "name": "sets list",
     "effects": "read_only",
     "examples": [
      "townsquare sets list"
     ],
     "output_kind": "data",
     "cardinality": "bounded"
    }
   },
   "post": {
    "tags": [
     "Sets"
    ],
    "summary": "Create group set",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WriteResult"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SetInput"
       }
      }
     }
    },
    "operationId": "create_group_set",
    "x-cli": {
     "name": "sets create",
     "effects": "non_idempotent",
     "examples": [
      "townsquare sets create --name 'All youth groups' --jids 120363000000000103@g.us,120363000000000104@g.us"
     ],
     "output_kind": "data",
     "cardinality": "single"
    }
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "bearer": {
    "type": "http",
    "scheme": "bearer",
    "description": "API key from Settings → API keys (tsq_...)"
   }
  },
  "schemas": {
...
   "Status": {
    "type": "object",
    "properties": {
     "connected": {
      "type": "boolean"
     },
     "phone": {
      "type": "string"
     },
     "timezone": {
      "type": "string"
     },
     "safe_mode": {
      "type": "boolean"
     },
     "allowlisted_targets": {
      "type": "integer"
     },
     "sent_today": {
      "type": "integer"
     },
     "daily_cap": {
      "type": "integer"
     },
     "undo": {
      "type": "string",
      "description": "What undo would revert next"
     },
     "redo": {
      "type": "string"
     },
     "now": {
      "type": "string",
      "format": "date-time"
     },
     "version": {
      "type": "string",
      "description": "App version, e.g. v0.5.0"
     },
     "commit": {
      "type": "string"
     },
     "key": {
      "type": "object",
      "properties": {
       "name": {
        "type": "string"
       },
       "scope": {
        "type": "string"
       }
      }
     },
     "demo": {
      "type": "boolean",
      "description": "Demo mode: sample data, nothing is ever sent"
     }
    }
   },
...
  },
  "responses": {
   "E400": {
    "description": "Bad request",
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    }
   },
   "E401": {
    "description": "Missing or invalid API key",
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    }
   },
...
  }
 }
}